Move new tests into separate file.
